Update DAC.c
made an update to the pin Mode, made it analog, updated the clock signal so it did DAC, fixed setting the value, and did not use bit 2 since it was by default it was set up and cleared with software. I also changed to set up the software trigger enable function in the CR register so it was set up with bits 3:5. will look for set up by software in the future.